Rename statistics app to avoid stdlib shadowing
