+++ backend/config/api_urls.py
@@ -48,5 +48,5 @@ urlpatterns = [
     path("content-migration/", include("content_migration.urls")),
     path("utils/", include("utils.urls")),
     path("ai-tracking/", include("ai_tracking.urls")),
-    path("statistics/", include("statistics.urls")),
+    path("statistics/", include("site_statistics.urls")),
 ]
