+++ backend/config/settings.py
@@ -203,7 +203,7 @@ LOCAL_APPS = [
     "content_migration",  # Background content migration system
     "utils",  # Utility features like value lists and schema system
     "ai_tracking",  # AI usage and cost tracking
-    "statistics",  # Integrated statistics and analytics system
+    "site_statistics",  # Integrated statistics and analytics system
 ]
 
 INSTALLED_APPS = DJANGO_APPS + THIRD_PARTY_APPS + LOCAL_APPS
@@ -606,11 +606,11 @@ CELERY_BEAT_SCHEDULE = {
         "schedule": 3600.0,  # Every hour
     },
     "aggregate-statistics-daily": {
-        "task": "statistics.tasks.aggregate_daily_stats",
+        "task": "site_statistics.tasks.aggregate_daily_stats",
         "schedule": 86400.0,  # Every 24 hours
     },
     "update-experiment-metrics-hourly": {
-        "task": "statistics.tasks.update_experiment_metrics",
+        "task": "site_statistics.tasks.update_experiment_metrics",
         "schedule": 3600.0,  # Every hour
     },
 }
