+++ backend/site_statistics/apps.py
@@ -3,5 +3,5 @@ from django.apps import AppConfig
 
 class StatisticsConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "statistics"
+    name = "site_statistics"
