+++ backend/site_statistics/management/commands/process_events.py
@@ -4,8 +4,8 @@ from django.core.management.base import BaseCommand
 from django.utils import timezone
 from django.utils.dateparse import parse_datetime
 
-from statistics.models import EventRaw
-from statistics.services.queue_driver import RabbitMqDriver
+from site_statistics.models import EventRaw
+from site_statistics.services.queue_driver import RabbitMqDriver
 from core.models import Tenant
 
 logger = logging.getLogger(__name__)
